:ambulance: Fix needs-update? and should-update?
The default should-update? was returning true if the props were
equal. If our function is pure, this is the exact opposite of what we
want. If the props have not changed there is no reason to update.

needs-update? was not updating if should-update? returned true. So in
order to make a component that always updated, you had to have a
should-update? false. That is definitely backwards.

After these changes, the changed always-update test passes as do all
the existing changes. These two mistakes in tandem made the
functionality work "properly" as long as you didn't provide a custom should-update?